refactor: [MEW-1937] extract numeric input handlers to useNumericInput composable
Move checkIfNumber and onPaste from AppEnterAmount into a shared
composable so any numeric input can reuse comma→period normalization.
